+++ report-service/src/main/java/com/example/reportservice/service/visit_auth/VisitAuthTxRequestService.java
@@ -1,7 +1,8 @@
-package com.example.reportservice.service;
+package com.example.reportservice.service.visit_auth;
 
 import com.example.reportservice.entity.VisitAuthRequest;
-import com.example.reportservice.repository.VisitAuthRequestRepository;
+import com.example.reportservice.repository.visit_auth.VisitAuthRequestRepository;
+import com.example.reportservice.service.outbox.OutBoxService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -12,9 +13,16 @@ import org.springframework.transaction.annotation.Transactional;
 public class VisitAuthTxRequestService {
 
     private final VisitAuthRequestRepository visitAuthRequestRepository;
+    private final OutBoxService messageService;
 
     @Transactional
     public void createVisitAuthRequest(final VisitAuthRequest visitAuthRequest) {
         visitAuthRequestRepository.save(visitAuthRequest);
+    }
+
+    @Transactional
+    public void authenticateVisitAuthRequest(VisitAuthRequest visitAuthRequest) {
+        visitAuthRequestRepository.authenticateVisitAuthRequest(visitAuthRequest.getId());
+        messageService.createMessage(visitAuthRequest);
     }
 }
